+++ hutool-core/src/test/java/cn/hutool/core/util/URLUtilTest.java
@@ -21,4 +21,15 @@ public class URLUtilTest {
 		normalize = URLUtil.normalize(url);
 		Assert.assertEquals("http://www.hutool.cn/aaa/bbb", normalize);
 	}
+	
+	@Test
+	public void normalizeTest2() {
+		String url = "http://www.hutool.cn//aaa/\\bbb?a=1&b=2";
+		String normalize = URLUtil.normalize(url);
+		Assert.assertEquals("http://www.hutool.cn/aaa/bbb?a=1&b=2", normalize);
+		
+		url = "www.hutool.cn//aaa/bbb?a=1&b=2";
+		normalize = URLUtil.normalize(url);
+		Assert.assertEquals("http://www.hutool.cn/aaa/bbb?a=1&b=2", normalize);
+	}
 }